Fix a bug related to yaml-cpp yaml-cpp parses uint8_t as char: https://github.com/jbeder/yaml-cpp/issues/201 This was causing python to read the wrong value in the object_type field of GoalHGMF. Solution: uint8 -> uint16

+++ src/comm_benchmarks.h
@@ -906,7 +906,7 @@ class BenchmarkHGMF
       goal_msg_->target_pose.y = goal_payload[1].as<double>();
       goal_msg_->target_pose.theta = 0.0;
 
-      goal_msg_->object_type = goal_payload[2].as<uint8_t>();
+      goal_msg_->object_type = goal_payload[2].as<uint16_t>();
 
       cout << "RECEIVING GOAL!!! " << endl;
